Refactor `ImageUtil` module, add `importModule` overrides
