+++ lib/util/Alert/index.js
@@ -1,3 +1,9 @@
+// If running outside the Scriptable environment...
+if (typeof config === 'undefined') {
+	// ... overwrite the `importModule()` global function.
+	global.importModule = modulePath => require(modulePath)
+}
+
 const { ErrorAlert } = importModule('./ErrorAlert');
 
 module.exports = {
