+++ lib/util/File/FileUtil.js
@@ -1,4 +1,4 @@
-const { ImageType } = importModule('../ImageUtil')
+const { ImageType } = importModule('../Image')
 
 const PERIOD = '.'
