+++ lib/util/File/FileUtil.test.js
@@ -1,10 +1,4 @@
-// If running outside the Scriptable environment...
-if (typeof config === 'undefined') {
-	// ... overwrite the `importModule()` global function.
-	global.importModule = modulePath => require(modulePath)
-}
-
-const { ImageType } = require('../ImageUtil')
+const { ImageType } = require('../Image')
 
 const { FileUtil } = require('./FileUtil')
