+++ lib/util/Image/ImageType.js
@@ -1,0 +1,9 @@
+class ImageType {
+
+	static JPG = 'jpg'
+	static PNG = 'png'
+}
+
+module.exports = {
+	ImageType,
+}
