+++ lib/util/Image/ImageUtil.js
@@ -1,3 +1,5 @@
+const { ImageType } = importModule('./ImageType')
+
 /**
  * Utility class for manipulating images.
  */
@@ -29,13 +31,6 @@ class ImageUtil {
 	}
 }
 
-class ImageType {
-
-	static JPG = 'jpg'
-	static PNG = 'png'
-}
-
 module.exports = {
-	ImageType,
 	ImageUtil,
 }
