+++ lib/util/Image/index.js
@@ -4,8 +4,10 @@ if (typeof config === 'undefined') {
 	global.importModule = modulePath => require(modulePath)
 }
 
-const { FileUtil } = importModule('./FileUtil')
+const { ImageType } = importModule('./ImageType')
+const { ImageUtil } = importModule('./ImageUtil')
 
 module.exports = {
-	FileUtil,
+	ImageType,
+	ImageUtil,
 }
